contact_book/requirements.txt
django-tastypie==0.14.0
MySQL-python==1.2.5
nose==1.3.7
django-nose==1.4.5
factory-boy==2.4.0
fake-factory==0.5.3
enum==0.4.6
mock==1.0.1
marisa_trie==0.7.4
elasticsearch>=5.0.0,<6.0.0
django-test-without-migrations==0.6
